tweaked types, need to set 0 crowdcount as NA instead of empty
